KEYCLOAK-943 Account service refactoring
