+++ server-spi-private/src/main/java/org/keycloak/broker/provider/util/SimpleHttp.java
@@ -17,16 +17,25 @@
 
 package org.keycloak.broker.provider.util;
 
-import org.apache.http.*;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.http.Header;
+import org.apache.http.HeaderIterator;
+import org.apache.http.HttpEntity;
+import org.apache.http.HttpResponse;
+import org.apache.http.NameValuePair;
+import org.apache.http.ParseException;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.entity.UrlEncodedFormEntity;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.methods.HttpRequestBase;
 import org.apache.http.client.utils.URIBuilder;
+import org.apache.http.entity.StringEntity;
 import org.apache.http.message.BasicNameValuePair;
 import org.keycloak.connections.httpclient.HttpClientProvider;
 import org.keycloak.models.KeycloakSession;
+import org.keycloak.util.JsonSerialization;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -47,25 +56,36 @@ import java.util.zip.GZIPInputStream;
  */
 public class SimpleHttp {
 
-    private KeycloakSession session;
+    private static final ObjectMapper mapper = new ObjectMapper();
+
+    private HttpClient client;
 
     private String url;
     private String method;
     private Map<String, String> headers;
     private Map<String, String> params;
+    private Object entity;
 
-    protected SimpleHttp(String url, String method, KeycloakSession session) {
-        this.session = session;
+    protected SimpleHttp(String url, String method, HttpClient client) {
+        this.client = client;
         this.url = url;
         this.method = method;
     }
 
     public static SimpleHttp doGet(String url, KeycloakSession session) {
-        return new SimpleHttp(url, "GET", session);
+        return doGet(url, session.getProvider(HttpClientProvider.class).getHttpClient());
+    }
+
+    public static SimpleHttp doGet(String url, HttpClient client) {
+        return new SimpleHttp(url, "GET", client);
     }
 
     public static SimpleHttp doPost(String url, KeycloakSession session) {
-        return new SimpleHttp(url, "POST", session);
+        return doPost(url, session.getProvider(HttpClientProvider.class).getHttpClient());
+    }
+
+    public static SimpleHttp doPost(String url, HttpClient client) {
+        return new SimpleHttp(url, "POST", client);
     }
 
     public SimpleHttp header(String name, String value) {
@@ -73,6 +93,11 @@ public class SimpleHttp {
             headers = new HashMap<String, String>();
         }
         headers.put(name, value);
+        return this;
+    }
+
+    public SimpleHttp json(Object entity) {
+        this.entity = entity;
         return this;
     }
 
@@ -84,10 +109,27 @@ public class SimpleHttp {
         return this;
     }
 
-    public String asString() throws IOException {
-        HttpClient httpClient = session.getProvider(HttpClientProvider.class).getHttpClient();
+    public SimpleHttp auth(String token) {
+        header("Authorization", "Bearer " + token);
+        return this;
+    }
 
-        HttpResponse response = makeRequest(httpClient);
+    public JsonNode asJson() throws IOException {
+        if (headers == null || !headers.containsKey("Accept")) {
+            header("Accept", "application/json");
+        }
+        return mapper.readTree(asString());
+    }
+
+    public <T> T asJson(Class<T> type) throws IOException {
+        if (headers == null || !headers.containsKey("Accept")) {
+            header("Accept", "application/json");
+        }
+        return JsonSerialization.readValue(asString(), type);
+    }
+
+    public String asString() throws IOException {
+        HttpResponse response = makeRequest();
 
         InputStream is;
         HttpEntity entity = response.getEntity();
@@ -113,14 +155,12 @@ public class SimpleHttp {
     }
 
     public int asStatus() throws IOException {
-        HttpClient httpClient = session.getProvider(HttpClientProvider.class).getHttpClient();
-
-        HttpResponse response = makeRequest(httpClient);
+        HttpResponse response = makeRequest();
 
         return response.getStatusLine().getStatusCode();
     }
 
-    private HttpResponse makeRequest(HttpClient httpClient) throws IOException {
+    private HttpResponse makeRequest() throws IOException {
         boolean get = method.equals("GET");
         boolean post = method.equals("POST");
 
@@ -130,7 +170,16 @@ public class SimpleHttp {
         }
 
         if (post) {
-            ((HttpPost) httpRequest).setEntity(getFormEntityFromParameter());
+            if (params != null) {
+                ((HttpPost) httpRequest).setEntity(getFormEntityFromParameter());
+            } else if (entity != null) {
+                if (headers == null || !headers.containsKey("Content-Type")) {
+                    header("Content-Type", "application/json");
+                }
+                ((HttpPost) httpRequest).setEntity(getJsonEntity());
+            } else {
+                throw new IllegalStateException("No content set");
+            }
         }
 
         if (headers != null) {
@@ -139,7 +188,7 @@ public class SimpleHttp {
             }
         }
 
-        return httpClient.execute(httpRequest);
+        return client.execute(httpRequest);
     }
 
     private URI appendParameterToUrl(String url) throws IOException {
@@ -161,12 +210,16 @@ public class SimpleHttp {
         return uri;
     }
 
+    private StringEntity getJsonEntity() throws IOException {
+        return new StringEntity(JsonSerialization.writeValueAsString(entity));
+    }
+
     private UrlEncodedFormEntity getFormEntityFromParameter() throws IOException{
         List<NameValuePair> urlParameters = new ArrayList<>();
 
         if (params != null) {
             for (Map.Entry<String, String> p : params.entrySet()) {
-                urlParameters.add(new BasicNameValuePair(p.getKey(), p.getValue()));
+                urlParameters. add(new BasicNameValuePair(p.getKey(), p.getValue()));
             }
         }
