+++ testsuite/integration-arquillian/tests/base/src/test/resources/testrealm.json
@@ -342,6 +342,13 @@
       ],
       "adminUrl": "http://localhost:8180/varnamedapp/base/admin",
       "secret": "password"
+    },
+    {
+      "clientId": "direct-grant",
+      "enabled": true,
+      "directAccessGrantsEnabled": true,
+      "secret": "password",
+      "webOrigins": [ "http://localtest.me:8180" ]
     }
   ],
   "roles" : {
